fix load_data
